Add working MD5 driver module
This commit adds a working MD5 computation procedure that requires
requested amount of bytes to process be aligned on a 64 byte
boundary. This requirement is there to keep the procedure simple
and potentially speed-optimal, depending on the caller to
partition its input into chunks of appropriate size and keep
calling the procedure until the very last chunk, which must be
padded (using another exported procedure of the module) before
being passed to the former procedure one final time.

The computation for arbitrary input which is loaded into the
WebAssembly memory and addressed through numeric offsets into said
memory, begins with a call to the [exported] "start" procedure
which initializes the so-called "computation context", followed by
at least one call to the "update" procedure, and finally one call
to the "pad" procedure followed by one final call to "update"
again. Upon return from the [final] "update" call, the MD5 of the
entire input (processed since the "start" call), will have been
computed and stored in the "context", addressable and retrievable
through the module's memory object.

Because of the constraints and properties described above, the
module is called a "driver" -- it will do the bulk of the work,
but the caller needs to properly chunk input, pad the last
piece of it and eventually load result from memory.