Add 16 variables to hold an entire data block
Since during a single block of 64 bytes worth of computation,
the 16 32-bit words are reused between the 4 different rounds,
each word is thus used 4 times. Loading each such 32-bit word
into a local variable once and then read the variable instead
during each step, is done with assumption that the cost of
loading of the word into a variable is outweighed by the benefit
that reading the variable (as opposed to loading from memory)
four times per block may bring with it. The performance cost or
benefit may of course depend on the WASM execution machine.