Document the module's procedures for posterity
MD5 isn't necessarily complicated, but in combination with typical stack machine syntax and semantics (WASM), I have felt on my own skin it's easy to forget how it all works, especially sometimes the crucial detail like the padding, and, again, looking at the companion repository that wraps use of the module in some higher-level JavaScript, didn't make it immediately obvious how the MD5 implementation works, which made me take the plunge, refresh the know-how I had when I originally implemented the module, and put it all in writing, to save whomever has to grok this later.

In short: more documentation can't hurt, right? As long as the single-source-of-truth still [largely] applies.